combine generated types
