+++ src/types/signature.ts
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-import { Envelope } from './bundle';
+import { Envelope } from './sigstore';
 import { OneOf } from './utility';
 
 interface VerificationMaterial {
