+++ src/types/sigstore/serialized.ts
