+++ store/trusted_root.json
@@ -1,4 +1,5 @@
 {
+    "mediaType": "application/vnd.dev.sigstore.trustedroot+json;version=0.1",
     "tlogs": [
       {
         "baseUrl": "https://tlog.sigstore.dev",
